combine coin fiat

+++ database/migrations/2024_11_21_085511_create_p2p_coin_fiat_pairs_table.php
@@ -17,7 +17,7 @@ class CreateP2pCoinFiatPairsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('coin_info_id');
             $table->unsignedBigInteger('fiat_info_id');
-            $table->unsignedTinyInteger("max_fiat_price");
+            $table->unsignedBigInteger("max_fiat_price");
             $table->foreign('coin_info_id')
                 ->references('id')
                 ->on('coin_infos');
